Daycare: auto-checkin, dog park arrivals, wake-before-send, XP tuning
- Auto-daycare on boot: reads last ROM's .sav from SD, checks in at 38s
- Immediate beacon + first event on auto-checkin (no 5-min wait)
- Dog park arrival events: type-matched Pokemon interact when new trainer appears
  (affinity = same type friendship, super-effective = instant rivalry)
- Wake-before-send: wake from light sleep 5s before beacon/event fires
  instead of keeping device always awake (fixes 5-min freeze from light sleep)
- No XP for rest/mischief events, only combat/social/explore
- Self-DM commands: messages to yourself treated as local commands
- "mmds" / "mmd s" shorthand for status
- Removed double-DM (events only sent from MonsterMeshModule, not PokemonDaycare)
- forceBeacon() and forceEvent() public methods with timer reset
- Eject cart freeze fix: full LVGL teardown on cart eject
- Beacon channel fix: accept PRIVATE_APP on primary channel
- Neighbor notification DM on new trainer discovery

+++ src/modules/monstermesh/MonsterMeshEmulator.h
@@ -24,7 +24,7 @@ static const uint16_t EMU_PALETTES[][4] = {
     { 0xFFFF, 0xAD55, 0x52AA, 0x0000 },  // 1: Grayscale
 };
 static constexpr int EMU_PALETTE_COUNT = sizeof(EMU_PALETTES) / sizeof(EMU_PALETTES[0]);
-extern volatile int g_emuPaletteIdx;
+extern volatile uint8_t g_emuPaletteIdx;
 
 // ── MonsterMesh Emulator ────────────────────────────────────────────────────────
 // Adapted from EmulatorApp. No longer owns TFT_eSPI directly — rendering goes
